requirements.txt
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
cloudinary==1.40.0
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3
